Split native sidebar DTO module
